+++ CHANGELOG.md
@@ -38,6 +38,8 @@
 - fix update unread badge on when muting / unmuting a chat #4020
 - fix update unread badge on receiving device messages #4020
 - fix target chat was not opened on notification click #3983
+- fix "Password and Account" dialog not indicating invalid credentials, making it seem that you can change password like this #4032
+- fix "Password and Account" not restoring original credentials on "Cancel" sometimes #4033
 
 <a id="1_46_1"></a>